+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-    Colonists now fill the cell in the manager job overview table.
+
 ## [0.10.0] - 2025-07-20
 
 ### Fixed